Update export functionality
Exporting to PDF via html was producing unpredictable results. I think this was due in part to the complexity of html used to generate rally grids, including highly nested elements, many class names and inline styles. Generating a json object and passing a small set of styles creates a consistent and clean output to pdf.